Organizing files and folder to follow atomic desing style
