+++ frontend/src/components/atoms/BorderedBox.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+import { Box } from '@mui/material';
+
+export const BorderedBox = styled(Box)`
+  display: flex;
+  padding: 2rem;
+  border: 1px solid black;
+  border-radius: 3%;
+`;
