+++ frontend/src/components/atoms/CenteredBox.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import { Box } from '@mui/material';
+
+export const CenteredBox = styled(Box)`
+  width: 100vw;
+	height: 100vh;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+`;
