+++ frontend/src/components/atoms/Form.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const Form = styled.form`
+	display: flex;
+	flex-direction: column;
+	row-gap: 1rem;
+`;
